Feature/IIA-2384: move convenience byte[] to JavaFX Image methods to KlFieldHelper
